Seperated helper functions into seperate keyboards file as suggested
