+++ quantum/usb_pids.json
@@ -1,0 +1,1 @@
+{"pids":{}}
